+++ luogu/P1423.cpp
@@ -18,10 +18,10 @@
 
 using namespace std;
 
-int main()
+long main()
 {
   double in, length = 2.0, ans = 0.0; // length 为当前步能游的距离，ans 为已游总距离
-  int step = 0;                       // step 为计量的步数
+  long step = 0;                       // step 为计量的步数
   cin >> in;
 
   for (; ans < in; ans += length, length *= 0.98, ++step)
