+++ luogu/P1601.cpp
@@ -21,7 +21,7 @@
 
 using namespace std;
 
-int main()
+long main()
 {
     string in1, in2;
     cin >> in1 >> in2;
@@ -36,8 +36,8 @@ int main()
         num1.swap(num2);
 
     vector<short> sum;
-    int temp = 0;
-    for (int p = 0; p < num1.size(); ++p)
+    long temp = 0;
+    for (long p = 0; p < num1.size(); ++p)
     {
         temp += num1.at(p) + num2.at(p);
         if (temp > 9)
@@ -51,7 +51,7 @@ int main()
             temp = 0;
         }
     }
-    for (int p = num1.size(); p < num2.size(); ++p)
+    for (long p = num1.size(); p < num2.size(); ++p)
     {
         temp += num2.at(p);
         if (temp > 9)
